refactor: split DiscountCalculator into SmeCalculator and LpCalculator subcomponents
